Other MER models/Original Nougat/nougat-latex-ocr/all_requirements.txt
transformers>=4.34.0
Pillow
python-Levenshtein
munch
anyconfig
albumentations==1.3.0
rapidfuzz==3.12.2
transformers==4.49.0
huggingface-hub==0.29.3
fsspec==2025.3.0
typing-extensions==4.12.2 
tifffile==2025.3.13
jiwer